+++ packages/connect/src/connector/specs/index.ts
@@ -1,18 +1,19 @@
-import { WellKnownChain } from "src/WellKnownChain"
-import ksmcc3 from "./js/ksmcc3.js"
-import polkadot from "./js/polkadot.js"
-import rococo_v2_2 from "./js/rococo_v2_2.js"
-import westend2 from "./js/westend2.js"
+import { WellKnownChain } from "../../WellKnownChain.js"
 
-const chains: Record<WellKnownChain, string> = {
-  ksmcc3,
-  polkadot,
-  rococo_v2_2,
-  westend2,
-}
+let chains: Record<WellKnownChain, Promise<{ chainSpec: string }>> | undefined
 
-export function getSpec(chain: string): string {
-  if (!Object.keys(chains).includes(chain))
+export async function getSpec(chain: string): Promise<string> {
+  if (!Object.keys(WellKnownChain).includes(chain))
     throw new Error("Invalid chain name")
-  return chains[chain as WellKnownChain]
+  if (!chains) {
+    chains = {
+      // Dynamic imports needs to be explicit for ParcelJS
+      // See https://github.com/parcel-bundler/parcel/issues/125
+      polkadot: import("./js/polkadot.js"),
+      ksmcc3: import("./js/ksmcc3.js"),
+      westend2: import("./js/westend2.js"),
+      rococo_v2_2: import("./js/rococo_v2_2.js"),
+    }
+  }
+  return (await chains[chain as WellKnownChain]).chainSpec
 }